change supportfile import to only reference node/browser compatible code
